+++ freppledb/common/dataload.py
@@ -190,7 +190,7 @@ def parseCSVdata(model, data, user=None, database=DEFAULT_DB_ALIAS, ping=False):
           # Argh... bool('0') returns True.
           return False
         else:
-          return val
+          return val if val != "" else None
       except KeyError as e:
         raise e
 
@@ -339,7 +339,7 @@ def _parseData(model, data, rowmapper, user, database, ping):
       if hasattr(model.objects, 'get_by_natural_key'):
         if model._meta.unique_together:
           natural_key = model._meta.unique_together[0]
-        elif hasattr(model, 'natural_key'):
+        elif hasattr(model, 'natural_key') and isinstance(model.natural_key, tuple):
           natural_key = model.natural_key
 
     # Case 2: Skip empty rows