+++ freppledb/execute/management/commands/loaddata.py
@@ -27,6 +27,7 @@ from django.template import Template, RequestContext
 from django.utils.translation import ugettext_lazy as _
 
 from freppledb.common.models import User
+from freppledb.common.middleware import _thread_locals
 from freppledb.execute.models import Task
 
 
@@ -139,6 +140,7 @@ class Command(loaddata.Command):
     now = datetime.now()
     task = None
     try:
+      setattr(_thread_locals, 'database', database)
       # Initialize the task
       if options['task']:
         try:
@@ -149,7 +151,8 @@ class Command(loaddata.Command):
           raise CommandError("Invalid task identifier")
         task.status = '0%'
         task.started = now
-        task.save(using=database, update_fields=['started', 'status'])
+        task.processid = os.getpid()
+        task.save(using=database, update_fields=['started', 'status', 'processid'])
       else:
         if options['user']:
           try:
@@ -162,6 +165,7 @@ class Command(loaddata.Command):
           name='loaddata', submitted=now, started=now, status='0%',
           user=user, arguments=' '.join(fixture_labels)
           )
+        task.processid = os.getpid()
         task.save(using=database)
 
       # Excecute the standard django command
@@ -201,6 +205,7 @@ class Command(loaddata.Command):
         # Task update
         task.status = 'Done'
         task.finished = datetime.now()
+        task.processid = None
         task.save(using=database, update_fields=['status', 'finished'])
 
     except Exception as e:
@@ -208,5 +213,9 @@ class Command(loaddata.Command):
         task.status = 'Failed'
         task.message = '%s' % e
         task.finished = datetime.now()
+        task.processid = None
         task.save(using=database, update_fields=['status', 'finished', 'message'])
       raise CommandError('%s' % e)
+
+    finally:
+      setattr(_thread_locals, 'database', None)