+++ freppledb/odoo/commands.py
@@ -299,17 +299,20 @@ class OdooWritePlan(PlanTask):
           yield '<operationplan id="%s" ordertype="PO" item=%s location=%s supplier=%s start="%s" end="%s" quantity="%s" location_id=%s item_id=%s criticality="%d"/>' % (
             i.id, quoteattr(i.item.name), quoteattr(i.location.name),
             quoteattr(i.supplier.name), i.start, i.end, i.quantity,
-            quoteattr(i.location.subcategory or ""), quoteattr(i.item.subcategory or ""),
+            quoteattr(i.location.subcategory), quoteattr(i.item.subcategory),
             int(i.criticality)
             )
         elif i.ordertype == "MO":
-          if not i.operation or not i.operation.source or not i.operation.source.startswith('odoo') or i.status not in ('proposed', 'approved'):
-            continue
+          if not i.operation or not i.operation.source \
+            or not i.operation.item \
+            or not i.operation.source.startswith('odoo') \
+            or i.status not in ('proposed', 'approved'):
+              continue
           cls.exported.append(i)
           yield '<operationplan id="%s" ordertype="MO" item=%s location=%s operation=%s start="%s" end="%s" quantity="%s" location_id=%s item_id=%s criticality="%d"/>' % (
             i.id, quoteattr(i.operation.item.name), quoteattr(i.operation.location.name),
             quoteattr(i.operation.name), i.start, i.end, i.quantity,
-            quoteattr(i.operation.location.subcategory or ""), quoteattr(i.operation.item.subcategory or ""),
+            quoteattr(i.operation.location.subcategory), quoteattr(i.operation.item.subcategory),
             int(i.criticality)
             )
       yield '</operationplans>'