+++ freppledb/odoo/migrations/0001_initial.py
@@ -19,15 +19,16 @@ from django.db import migrations
 
 
 def loadParameters(apps, schema_editor):
+  from django.core.management.commands.loaddata import Command
   call_command(
-    'loaddata', "parameters.json", app_label="odoo",
+    Command(), "parameters.json", app_label="odoo",
     verbosity=0, database=schema_editor.connection.alias
     )
 
 
 class Migration(migrations.Migration):
   dependencies = [
-      ('common', '0001_initial'),
+    ('common', '0001_initial')
   ]
 
   operations = [